requirements.txt
articles==0.1.0
asgiref==3.6.0
category==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-allauth==0.52.0
django-braces==1.15.0
django-extensions==3.2.1
django-jazzmin==2.6.0
django-oauth-toolkit==2.2.0
django-phone-field==1.8.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
dnspython==2.3.0
fontawesomefree==6.2.1
ggg==0.0.1
idna==3.4
jwcrypto==1.4.2
Markdown==3.4.1
mongoengine==0.26.0
numpy==1.24.1
oauthlib==3.2.2
phonenumbers==8.13.5
Pillow==9.3.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rust_category==0.2.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.3.0
social-auth-storage-mongoengine==1.0.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
